Refactor parser package
