+++ pkg/logs/config/source.go
@@ -33,7 +33,7 @@ func NewLogSource(name string, config *LogsConfig) *LogSource {
 		inputs:   make(map[string]bool),
 		lock:     &sync.Mutex{},
 		Messages: NewMessages(),
-		Parser:   parser.NewIdentityParser(),
+		Parser:   parser.NewNoopParser(),
 	}
 }
 
@@ -62,7 +62,9 @@ func (s *LogSource) GetInputs() []string {
 	return inputs
 }
 
-// AddParser add a parser
-func (s *LogSource) AddParser(pm parser.Parser) {
-	s.Parser = pm
+// SetParser adds a parser
+func (s *LogSource) SetParser(parser parser.Parser) {
+	s.lock.Lock()
+	s.Parser = parser
+	s.lock.Unlock()
 }
