+++ pkg/logs/decoder/line_handler_test.go
@@ -24,7 +24,6 @@ type MockParser struct {
 	header []byte
 }
 
-// parser.NewIdentityParser returns a new Unwrapper
 func NewMockParser(header string) parser.Parser {
 	return &MockParser{header: []byte(header)}
 }
@@ -40,7 +39,7 @@ func (u *MockParser) Unwrap(line []byte) ([]byte, error) {
 
 func TestSingleLineHandler(t *testing.T) {
 	outputChan := make(chan *Output, 10)
-	h := NewSingleLineHandler(outputChan, parser.NewIdentityParser())
+	h := NewSingleLineHandler(outputChan, parser.NewNoopParser())
 	h.Start()
 
 	var output *Output
@@ -81,7 +80,7 @@ func TestSingleLineHandler(t *testing.T) {
 
 func TestTrimSingleLine(t *testing.T) {
 	outputChan := make(chan *Output, 10)
-	h := NewSingleLineHandler(outputChan, parser.NewIdentityParser())
+	h := NewSingleLineHandler(outputChan, parser.NewNoopParser())
 	h.Start()
 
 	var output *Output
@@ -100,7 +99,7 @@ func TestTrimSingleLine(t *testing.T) {
 func TestMultiLineHandler(t *testing.T) {
 	re := regexp.MustCompile("[0-9]+\\.")
 	outputChan := make(chan *Output, 10)
-	h := NewMultiLineHandler(outputChan, re, 10*time.Millisecond, parser.NewIdentityParser())
+	h := NewMultiLineHandler(outputChan, re, 10*time.Millisecond, parser.NewNoopParser())
 	h.Start()
 
 	var output *Output
@@ -153,7 +152,7 @@ func TestMultiLineHandler(t *testing.T) {
 func TestTrimMultiLine(t *testing.T) {
 	re := regexp.MustCompile("[0-9]+\\.")
 	outputChan := make(chan *Output, 10)
-	h := NewMultiLineHandler(outputChan, re, 10*time.Millisecond, parser.NewIdentityParser())
+	h := NewMultiLineHandler(outputChan, re, 10*time.Millisecond, parser.NewNoopParser())
 	h.Start()
 
 	var output *Output
